+++ dotNET/PdfClown.Samples/Samples/TextFrameSample.cs
@@ -43,7 +43,7 @@ namespace PdfClown.Samples.CLI
 
             colors::Color textColor = new colors::DeviceRGBColor(115 / 255d, 164 / 255d, 232 / 255d);
             composer.SetFillColor(textColor);
-            composer.SetLineDash(new LineDash(new double[] { 10 }));
+            composer.SetLineDash(new LineDash(new float[] { 10 }));
             composer.SetLineWidth(.25);
 
             BlockComposer blockComposer = new BlockComposer(composer);
