+++ dotNET/PdfClown.Samples/Samples/TextInfoExtractionSample.cs
@@ -106,7 +106,7 @@ namespace PdfClown.Samples.CLI
 
                         // Drawing text string bounding box...
                         composer.BeginLocalState();
-                        composer.SetLineDash(new LineDash(new double[] { 5 }));
+                        composer.SetLineDash(new LineDash(new float[] { 5 }));
                         composer.SetStrokeColor(textStringBoxColor);
                         composer.DrawRectangle(textString.Box.Value);
                         composer.Stroke();
